Function rustyknife::rfc3461::orcpt_address

pub fn orcpt_address(
    input: &[u8]
) -> NomResult<'_, (Cow<'_, str>, Cow<'_, str>)>

Parse the ESMTP ORCPT parameter that may be present on a RCPT TO command.

Returns the address type and the decoded original recipient address.

Examples

use rustyknife::rfc3461::orcpt_address;

let (_, split) = orcpt_address(b"rfc822;bob@example.org").unwrap();

assert_eq!(split, ("rfc822".into(), "bob@example.org".into()));