Module rustyknife::rfc5321

Parsers for SMTP command syntax

Structs

Keyword
ESMTP parameter keyword.

Param
ESMTP parameter.

Params
Newtype over a slice of Param for display purposes.

Path
Path with source route.

SMTPString
A generic SMTP string built from an atom or a quoted string

Value
ESMTP parameter value.

Enums

Command
The base SMTP command set

ForwardPath
Represents a forward path from the "RCPT TO" command.

ReversePath
Represents a reverse path from the "MAIL FROM" command.

Traits

UTF8Policy

Functions

bdat_command
Parse a BDAT command from RFC 3030

command
Parse any basic SMTP command.

data_command
Parse an SMTP DATA command.

ehlo_command
Parse an SMTP EHLO command.

expn_command
Parse an SMTP EXPN command.

helo_command
Parse an SMTP HELO command.

help_command
Parse an SMTP HELP command.

mail_command
Parse an SMTP MAIL FROM command.

noop_command
Parse an SMTP NOOP command.

quit_command
Parse an SMTP QUIT command.

rcpt_command
Parse an SMTP RCPT TO command.

rset_command
Parse an SMTP RSET command.

starttls_command
Parse a STARTTLS command from RFC 3207

validate_address
Validates an email address.

vrfy_command
Parse an SMTP VRFY command.