Module rustyknife::types

Types shared by SMTP and Internet Message Format

IMF allows comments and folding spaces in many places but is very close to the SMTP syntax.

If glaring incompatibilites are found in practice, the crate::rfc5321 and crate::rfc5322 modules will get their own variants of these types.

Structs such as [types::Domain] and [types::QuotedString] are newtypes around String to make sure they can only be constructed from valid values.

Structs

Domain
A domain name such as used by DNS.

DotAtom
A string consisting of atoms separated by periods.

Mailbox
A valid email address.

QuotedString
A quoted string that may contain spaces.

Enums

AddressLiteral
A network address literal.

DomainPart
The domain part of an address following the "@" in an email address.

LocalPart
The local part of an address preceding the "@" in an email address.